requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.2
Pillow>=9.0.0
PyPDF2>=3.0.0
torch
numpy
tqdm